Improve keyword detection in `TSModuleDeclaration`
