+++ src/language-js/print/typescript.js
@@ -8,6 +8,7 @@ import {
 } from "../../document/builders.js";
 import UnexpectedNodeError from "../../utils/unexpected-node-error.js";
 import { locStart } from "../loc.js";
+import getTextWithoutComments from "../utils/get-text-without-comments.js";
 import {
   isArrayOrTupleExpression,
   isObjectOrRecordExpression,
@@ -311,9 +312,9 @@ function printTypescript(path, options, print) {
             node.kind ??
             // TODO: Use `node.kind` when babel update AST
             (isStringLiteral(node.id) ||
-            /(?:^|\s)module(?:\s|$)/.test(
-              options.originalText.slice(locStart(node), locStart(node.id)),
-            )
+            getTextWithoutComments(options, locStart(node), locStart(node.id))
+              .trim()
+              .endsWith("module")
               ? "module"
               : "namespace");
           parts.push(kind, " ");
