+++ tests/format/typescript/module/keyword.ts
@@ -29,3 +29,7 @@ namespace X {
     const x = 1;
   }
 }
+
+namespace /* module */ X {}
+module /* namespace */ X {}
+module /* namespace */ "x" {}
